[FLEX-94]: Added bets update to statscore service data fetch process.

+++ src/main/java/io/limeup/flexbets/sport/dto/statscore/StatScoreCompetitionDTO.java
@@ -29,5 +29,5 @@ public class StatScoreCompetitionDTO {
     private String status;
     private List<StatScoreSeasonDTO> seasons;
     private StatScoreSeasonDTO season;
-    private Long lsId;
+    private Integer lsId;
 }
