+++ src/main/java/io/limeup/flexbets/sport/model/Competition.java
@@ -52,5 +52,5 @@ public class Competition {
     private String gender;
 
     @Column(name = "ls_id", unique = true)
-    private Long lsId;
+    private Integer lsId;
 }
